+++ maud/tests/basic_syntax.rs
@@ -52,12 +52,6 @@ fn empty_elements() {
 }
 
 #[test]
-fn empty_elements_slash() {
-    let result = html! { "pinkie" br / "pie" };
-    assert_eq!(result.into_string(), "pinkie<br>pie");
-}
-
-#[test]
 fn simple_attributes() {
     let result = html! {
         link rel="stylesheet" href="styles.css";
@@ -116,6 +110,47 @@ fn toggle_empty_attributes_braces() {
 fn empty_attributes_question_mark() {
     let result = html! { input checked? disabled?[true]; };
     assert_eq!(result.into_string(), "<input checked disabled>");
+}
+
+#[test]
+fn optional_attribute_some() {
+    let result = html! { input value=[Some("value")]; };
+    assert_eq!(result.into_string(), r#"<input value="value">"#);
+}
+
+#[test]
+fn optional_attribute_none() {
+    let result = html! { input value=[None as Option<&str>]; };
+    assert_eq!(result.into_string(), r#"<input>"#);
+}
+
+#[test]
+fn optional_attribute_non_string_some() {
+    let result = html! { input value=[Some(42)]; };
+    assert_eq!(result.into_string(), r#"<input value="42">"#);
+}
+
+#[test]
+fn optional_attribute_variable() {
+    let x = Some(42);
+    let result = html! { input value=[x]; };
+    assert_eq!(result.into_string(), r#"<input value="42">"#);
+}
+
+#[test]
+fn optional_attribute_inner_value_evaluated_only_once() {
+    let mut count = 0;
+    html! { input value=[{ count += 1; Some("picklebarrelkumquat") }]; };
+    assert_eq!(count, 1);
+}
+
+#[test]
+fn optional_attribute_braces() {
+    struct Pony {
+        cuteness: Option<i32>,
+    }
+    let result = html! { input value=[Pony { cuteness: Some(9000) }.cuteness]; };
+    assert_eq!(result.into_string(), r#"<input value="9000">"#);
 }
 
 #[test]
@@ -248,7 +283,7 @@ fn mixed_classes() {
 
 #[test]
 fn id_shorthand() {
-    let result = html! { p { "Hi, " span#thing { "Lyra" } "!" } };
+    let result = html! { p { "Hi, " span #thing { "Lyra" } "!" } };
     assert_eq!(
         result.into_string(),
         r#"<p>Hi, <span id="thing">Lyra</span>!</p>"#
@@ -257,7 +292,7 @@ fn id_shorthand() {
 
 #[test]
 fn id_string() {
-    let result = html! { h1#"pinkie-123" { "Pinkie Pie" } };
+    let result = html! { h1 #"pinkie-123" { "Pinkie Pie" } };
     assert_eq!(
         result.into_string(),
         r#"<h1 id="pinkie-123">Pinkie Pie</h1>"#