+++ maud/tests/warnings/void-element-slash.rs
@@ -1,0 +1,9 @@
+use maud::html;
+
+fn main() {
+    html! {
+        br /
+        // Make sure we're not stopping on the first error
+        input type="text" /
+    };
+}