+++ maud_macros/src/lib.rs
@@ -1,4 +1,4 @@
-#![doc(html_root_url = "https://docs.rs/maud_macros/0.22.3")]
+#![doc(html_root_url = "https://docs.rs/maud_macros/0.24.0")]
 // TokenStream values are reference counted, and the mental overhead of tracking
 // lifetimes outweighs the marginal gains from explicit borrowing
 #![allow(clippy::needless_pass_by_value)]
@@ -6,6 +6,7 @@
 extern crate proc_macro;
 
 mod ast;
+mod escape;
 mod generate;
 mod parse;
 
@@ -17,14 +18,6 @@ use quote::quote;
 #[proc_macro_error]
 pub fn html(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
     expand(input.into()).into()
-}
-
-#[proc_macro]
-#[proc_macro_error]
-pub fn html_debug(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
-    let expr = expand(input.into());
-    println!("expansion:\n{}", expr);
-    expr.into()
 }
 
 fn expand(input: TokenStream) -> TokenStream {