+++ maud_macros/src/parse.rs
@@ -538,6 +538,14 @@ impl Parser {
             {
                 // Void element
                 self.advance();
+                if punct.as_char() == '/' {
+                    emit_error!(
+                        punct,
+                        "void elements must use `;`, not `/`";
+                        help = "change this to `;`";
+                        help = "see https://github.com/lambda-fairy/maud/pull/315 for details";
+                    );
+                }
                 ast::ElementBody::Void {
                     semi_span: SpanRange::single_span(punct.span()),
                 }
@@ -571,21 +579,24 @@ impl Parser {
                         // Parse a value under an attribute context
                         assert!(self.current_attr.is_none());
                         self.current_attr = Some(ast::name_to_string(name.clone()));
-                        let value = self.markup();
+                        let attr_type = match self.attr_toggler() {
+                            Some(toggler) => ast::AttrType::Optional { toggler },
+                            None => {
+                                let value = self.markup();
+                                ast::AttrType::Normal { value }
+                            }
+                        };
                         self.current_attr = None;
-                        attrs.push(ast::Attr::Attribute {
-                            attribute: ast::Attribute {
-                                name,
-                                attr_type: ast::AttrType::Normal { value },
-                            },
+                        attrs.push(ast::Attr::Named {
+                            named_attr: ast::NamedAttr { name, attr_type },
                         });
                     }
                     // Empty attribute (legacy syntax)
                     Some(TokenTree::Punct(ref punct)) if punct.as_char() == '?' => {
                         self.advance();
                         let toggler = self.attr_toggler();
-                        attrs.push(ast::Attr::Attribute {
-                            attribute: ast::Attribute {
+                        attrs.push(ast::Attr::Named {
+                            named_attr: ast::NamedAttr {
                                 name: name.clone(),
                                 attr_type: ast::AttrType::Empty { toggler },
                             },
@@ -594,8 +605,8 @@ impl Parser {
                     // Empty attribute (new syntax)
                     _ => {
                         let toggler = self.attr_toggler();
-                        attrs.push(ast::Attr::Attribute {
-                            attribute: ast::Attribute {
+                        attrs.push(ast::Attr::Named {
+                            named_attr: ast::NamedAttr {
                                 name: name.clone(),
                                 attr_type: ast::AttrType::Empty { toggler },
                             },
@@ -643,7 +654,7 @@ impl Parser {
                     "class".to_string()
                 }
                 ast::Attr::Id { .. } => "id".to_string(),
-                ast::Attr::Attribute { attribute } => attribute
+                ast::Attr::Named { named_attr } => named_attr
                     .name
                     .clone()
                     .into_iter()